requirements.txt
streamlit
pyheif
img2pdf
Pillow
reportlab
PyPDF2